[all] Improved rubygems citizenship
Removed a lot of the complexity in specifying
require paths and which specs to run, relying
on defaults and not assuming rubygems as much as
is practical.

It's still possible to either run 'rake spec' or
run the 'spec' command on individual files. If
you want to run specs for individual files, you
will need to do something to the effect of:

  export RUBYOPT=-Ispec

Also, when running from a typical dev setup
(dm-core and dm-more are in the same folder),
specs will pick up the local version of dm-core
instead of the gem version. This also happens
for dm-more plugin specs that depend on other
dm-more plugins.