Imported migration related code from dm-core
This means that the automigration features are now
only available if dm-migrations along with at least
one of the dm-xxx-adapter gems that support that
feature, are required.

Also, dm-migrations now holds the automigration
related code for every adapter that supports the
concept of automigration. This means that if
adapter authors properly want to support this
feature for their adapters, the code has to be
put into dm-migrations.